sdk - retrieve_tx_on_network refactor
