+++ framework/snippets/src/multi/interactor_multi_sc_process.rs
@@ -1,6 +1,8 @@
 use crate::sdk::data::transaction::{Transaction, TransactionOnNetwork};
 use crate::{multiversx_sc::types::Address, Interactor, Sender};
 use futures::future::join_all;
+use multiversx_sdk::gateway::SendTxRequest;
+use multiversx_sdk::retrieve_tx_on_network;
 use std::collections::HashSet;
 
 pub(crate) type Txs = Vec<Transaction>;
@@ -26,12 +28,12 @@ impl Interactor {
         for tx in &txs {
             let tx_hash = self
                 .proxy
-                .send_transaction(tx)
+                .http_request(SendTxRequest(tx))
                 .await
                 .expect("failed to send transaction");
 
             println!("process tx hash: {tx_hash} with nonce: {}", tx.nonce);
-            futures.push(self.proxy.retrieve_tx_on_network(tx_hash.clone()));
+            futures.push(retrieve_tx_on_network(&self.proxy, tx_hash.clone()));
         }
 
         self.generate_blocks(4).await.unwrap();
