+++ sdk/core/src/lib.rs
@@ -1,6 +1,9 @@
 pub mod crypto;
 pub mod data;
 pub mod gateway;
+mod retrieve_tx_on_network;
 pub mod test_wallets;
 pub mod utils;
 pub mod wallet;
+
+pub use retrieve_tx_on_network::retrieve_tx_on_network;
